+++ src/index.ts
@@ -4,3 +4,5 @@ export * from './abstractClient';
 export * from './licenses/';
 export * from './general/';
 export * from './subscriptions/';
+export * from './customers/';
+export * from './getResult';